requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
astroid==2.12.10
attrs==22.1.0
autoflake==1.7.0
autopep8==1.7.0
bcrypt==3.2.2
bioblend==0.17.0
black==22.3.0
bleach==5.0.0
boltons==21.0.0
boto==2.49.0
build==0.8.0
certifi==2022.6.15
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
cli-test-helpers==3.1.0
click==8.1.3
config==0.5.1
confusable-homoglyphs==3.2.0
coverage==6.5.0
cryptography==37.0.4
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.4
docutils==0.18.1
ephemeris==0.10.7
exceptiongroup==1.0.1
fastcore==1.5.14
filelock==3.4.2
galaxy-containers==21.9.0
galaxy-lib==19.5.2
galaxy-tool-util==21.9.2
galaxy-util==21.9.0
ghapi==1.0.0
identify==2.5.6
idna==3.3
importlib==1.0.4
importlib-resources==5.9.0
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
mkdir==2020.12.3
mypy-extensions==0.4.3
nodeenv==1.7.0
nose==1.3.7
orderdict==2020.12.3
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
pep517==0.13.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
promise==2.3
protobuf==3.19.5
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
pyflakes==2.5.0
PyGithub==1.55
PyJWT==2.4.0
pylint==2.15.3
PyNaCl==1.5.0
pyparsing==3.0.9
pysam==0.19.1
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
PyYAML==6.0
repoze.lru==0.7
requests==2.28.0
requests-toolbelt==0.9.1
Routes==2.5.1
scp==0.14.4
setupcfg==2020.12.3
setuppy-generator==2020.12.2
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
text-unidecode==1.3
tinytag==1.8.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
toposort==1.7
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
values==2020.12.3
virtualenv==20.13.1
webencodings==0.5.1
whitenoise==6.0.0
-e git+ssh://git@github.com/combat-sars-cov-2/okapi.git@116187f0a51560648289c7a65576c013b9530d95#egg=workbench_cli
wrapt==1.14.1
write==2020.12.3
zipp==3.8.1
zipstream-new==1.1.8
